+++ index.js
@@ -9,5 +9,7 @@ module.exports = {
     oneLevel: require('./one-level.json'),
     twoLevels: require('./two-levels.json'),
     threeLevels: require('./three-levels.json'),
-    threeLevelsFull: require('./three-levels-full.json')
+    threeLevelsFull: require('./three-levels-full.json'),
+    threeLevelsMap: require('./three-levels-map.json'),
+    threeLevelsMapFull: require('./three-levels-map-full.json')
 };