+++ scripts/one-level-map.js
@@ -1,0 +1,17 @@
+
+'use strict';
+
+var _ = require('lodash');
+
+var province = require('../province.json');
+var makeJson = require('../util/make-json');
+var logger = require('../util/logger');
+
+var result = {};
+
+_.forEach(province, (name, code) => {
+    result[code] = {code: code, name: name};
+});
+
+logger.success(`一共 ${_.keys(result).length} 个省级行政单位`);
+makeJson('one-level-map', result);