+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # china-regions changelog
 
+## 0.0.2
+
+- add `one-level-map.json`
+- add `two-levels-map.json`
+- add `three-levels-map.json`
+- add `three-levels-map-full.json`
+
 ## 0.0.1
 
 - update to `data_location#73338af`